Move exercise-toasters logic to dedicated directive
