+++ angular.json
@@ -13,6 +13,10 @@
           "style": "scss",
           "skipTests": true,
           "standalone": true
+        },
+        "@schematics/angular:directive": {
+          "skipTests": true,
+          "standalone": true
         }
       },
       "architect": {
